requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.46
databases[aiosqlite]==0.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
coverage==7.3.2
